+++ packages/frontend-v2/components/Plan/ScheduleYear.tsx
@@ -93,7 +93,7 @@ export const ScheduleYear: React.FC<ScheduleYearProps> = ({
       {isExpanded && (
         <Grid templateColumns="repeat(4, 1fr)" minHeight="220px">
           <ScheduleTerm
-            scheduleYear={scheduleYear}
+            yearNum={scheduleYear.year}
             termCoReqErr={yearCoReqError?.fall}
             termPreReqErr={yearPreReqError?.fall}
             scheduleTerm={scheduleYear.fall}
@@ -101,7 +101,7 @@ export const ScheduleYear: React.FC<ScheduleYearProps> = ({
             removeCourseFromTermInCurrPlan={removeCourseFromTermInCurrPlan}
           />
           <ScheduleTerm
-            scheduleYear={scheduleYear}
+            yearNum={scheduleYear.year}
             termCoReqErr={yearCoReqError?.spring}
             termPreReqErr={yearPreReqError?.spring}
             scheduleTerm={scheduleYear.spring}
@@ -110,7 +110,7 @@ export const ScheduleYear: React.FC<ScheduleYearProps> = ({
           />
           {/* TODO: support summer full term */}
           <ScheduleTerm
-            scheduleYear={scheduleYear}
+            yearNum={scheduleYear.year}
             termCoReqErr={yearCoReqError?.summer1}
             termPreReqErr={yearPreReqError?.summer1}
             scheduleTerm={scheduleYear.summer1}
@@ -118,7 +118,7 @@ export const ScheduleYear: React.FC<ScheduleYearProps> = ({
             removeCourseFromTermInCurrPlan={removeCourseFromTermInCurrPlan}
           />
           <ScheduleTerm
-            scheduleYear={scheduleYear}
+            yearNum={scheduleYear.year}
             termCoReqErr={yearCoReqError?.summer2}
             termPreReqErr={yearPreReqError?.summer2}
             scheduleTerm={scheduleYear.summer2}
