+++ packages/frontend-v2/utils/plan/updatePlanOnDragEnd.ts
@@ -4,14 +4,12 @@ import {
   Schedule2,
   ScheduleCourse2,
   ScheduleTerm2,
-  SeasonEnum,
 } from "@graduate/common";
 import produce from "immer";
 import { toast } from "react-toastify";
 import { getCourseDisplayString } from "../course";
 import { getSeasonDisplayWord } from "./getSeasonDisplayWord";
 import { isCourseInTerm } from "./isCourseInTerm";
-import { logger } from "../logger";
 
 /**
  * Updates the schedule of plan when a course is dragged from one term to
@@ -45,10 +43,7 @@ export const updatePlanOnDragEnd = (
       throw new Error("Term of the course that is dragged over isn't found");
     }
 
-    console.log(plan);
-
     const year = plan.schedule.years.find((year) => {
-      console.log(year);
       const res = Object.values(year).find((term) => term.id === newTerm.id);
       return res;
     });
@@ -132,43 +127,4 @@ export const flattenScheduleToTerms = <T>(schedule: Schedule2<T>) => {
   );
 
   return scheduleTerms;
-};
-
-/** Finds a current term and season from a year and plan */
-export const findTerm = (
-  termSeason: SeasonEnum,
-  plan: PlanModel<string>,
-  termYear: number
-) => {
-  const scheduleYear = plan.schedule.years.find(
-    (year) => termYear === year.year
-  );
-
-  if (!scheduleYear) {
-    const errMsg = "Term with given year and season not found.";
-    logger.debug("UpdatePlanOnDragEnd.ts", errMsg, termYear, termSeason, plan);
-    throw new Error("Term with given year and season not found.");
-  }
-
-  let term = undefined;
-
-  switch (termSeason) {
-    case SeasonEnum.FL:
-      term = scheduleYear.fall;
-      break;
-    case SeasonEnum.SP:
-      term = scheduleYear.spring;
-      break;
-    case SeasonEnum.S1:
-      term = scheduleYear.summer1;
-      break;
-    case SeasonEnum.S2:
-      term = scheduleYear.summer2;
-      break;
-    case SeasonEnum.SM:
-      throw new Error(
-        "Full summer doesn't exist right now come back when it exists!"
-      );
-  }
-  return term;
 };
